optional priority
git-svn-id: http://fritzing.googlecode.com/svn/trunk@6814 85f94c43-f63a-0410-9d22-21439860c375

+++ fritzing/src/partsbinpalette/graphicsflowlayout.cpp
@@ -80,7 +80,7 @@ int GraphicsFlowLayout::doLayout(const QRectF &rect) {
 
 		x = nextX;
 		// item->preferredSize().height() returns qreal, armel compiler complains
-		lineHeight = qMax(lineHeight, (double) item->preferredSize().height());
+		lineHeight = qMax(lineHeight, item->preferredSize().height());
 	}
 
 	m_lastWidth = rect.width();
